+++ kakaowork/__init__.py
@@ -1,14 +1,57 @@
 # noqa: D104
-from kakaowork.blockkit import (BlockType, ButtonStyle, ButtonActionType, HeaderStyle, BlockKitType, SelectBlockOption, TextBlock, ImageLinkBlock, ButtonBlock,
-                                DividerBlock, HeaderBlock, ActionBlock, DescriptionBlock, SectionBlock, ContextBlock, LabelBlock, InputBlock, SelectBlock,
-                                BlockKitBuilder)
+from kakaowork.blockkit import (
+    BlockType,
+    ButtonStyle,
+    ButtonActionType,
+    HeaderStyle,
+    BlockKitType,
+    SelectBlockOption,
+    TextInlineType,
+    TextInlineColor,
+    TextInline,
+    TextBlock,
+    ImageLinkBlock,
+    ButtonBlock,
+    DividerBlock,
+    HeaderBlock,
+    ActionBlock,
+    DescriptionBlock,
+    SectionBlock,
+    ContextBlock,
+    LabelBlock,
+    InputBlock,
+    SelectBlock,
+    BlockKitBuilder,
+)
 
 from kakaowork.exceptions import (KakaoworkError, InvalidBlock, InvalidBlockType, NoValueError)
 
 from kakaowork.client import (Kakaowork, AsyncKakaowork)
 
-from kakaowork.models import (ErrorCode, ConversationType, ColorTone, ProfileNameFormat, ProfilePositionFormat, BotStatus, ErrorField, UserIdentificationField,
-                              UserField, ConversationField, MessageField, DepartmentField, SpaceField, BotField, BaseResponse, UserResponse, UserListResponse,
-                              ConversationResponse, ConversationListResponse, MessageResponse, DepartmentListResponse, SpaceResponse, BotResponse)
+from kakaowork.models import (
+    ErrorCode,
+    ConversationType,
+    ColorTone,
+    ProfileNameFormat,
+    ProfilePositionFormat,
+    BotStatus,
+    ErrorField,
+    UserIdentificationField,
+    UserField,
+    ConversationField,
+    MessageField,
+    DepartmentField,
+    SpaceField,
+    BotField,
+    BaseResponse,
+    UserResponse,
+    UserListResponse,
+    ConversationResponse,
+    ConversationListResponse,
+    MessageResponse,
+    DepartmentListResponse,
+    SpaceResponse,
+    BotResponse,
+)
 
-__version__ = '0.4.1'
+__version__ = '0.5.0'